requirements.txt
langgraph>=0.1.0
langchain-community>=0.0.20
langchain-ollama>=0.1.0
ollama>=0.2.0
chromadb>=0.4.0
sentence-transformers>=2.7.0
umap-learn>=0.5.6
networkx>=3.2
pyvis>=0.3.2
streamlit>=1.35.0
plotly>=5.20.0
pyyaml>=6.0
numpy>=1.26.0
scikit-learn>=1.4.0
pandas>=2.1.0
duckduckgo-search>=6.0.0
